pyproject.toml
[tool.poetry]
name = "eurotek-employee"
version = "0.1.0"
description = ""
authors = ["vadjango <z9661744@gmail.com>"]
readme = "README.md"
packages = [{ include = "eurotek_employee" }]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.3"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
django-phonenumber-field = "^7.1.0"
phonenumbers = "^8.13.17"
pillow = "^10.0.0"
django-extensions = "^3.2.3"
ipython = "^8.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-trench = "^0.3.1"
django-configurations = "^2.4.1"
pytest-django = "^4.5.2"
drf-nested-routers = "^0.93.4"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.1.0"
pytest-asyncio = "^0.21.1"
django-cors-headers = "^4.2.0"
celery = "^5.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
